requirements.txt
django
pillow
django-admin-interface
django-rest-framework
social-auth-app-django
django-pwa
django-crispy-forms==1.13.0
crispy-tailwind==0.5.0